+++ testutil/configurator/configurator.go
@@ -260,7 +260,7 @@ func MintModule() ModuleOption {
 			GolangBindings: []*appv1alpha1.GolangBinding{
 				{
 					InterfaceType:  "github.com/cosmos/cosmos-sdk/x/mint/types/types.StakingKeeper",
-					Implementation: "github.com/cosmos/cosmos-sdk/x/staking/keeper/*keeper.Keeper",
+					Implementation: "cosmossdk.io/x/staking/keeper/*keeper.Keeper",
 				},
 			},
 		}