+++ x/authz/testutil/app_config.go
@@ -1,6 +1,9 @@
 package testutil
 
 import (
+	_ "cosmossdk.io/x/gov"     // import as blank for app wiring
+	_ "cosmossdk.io/x/staking" // import as blank for app wiring
+
 	"github.com/cosmos/cosmos-sdk/testutil/configurator"
 	_ "github.com/cosmos/cosmos-sdk/x/auth"           // import as blank for app wiring
 	_ "github.com/cosmos/cosmos-sdk/x/auth/tx/config" // import as blank for app wiring
@@ -8,9 +11,7 @@ import (
 	_ "github.com/cosmos/cosmos-sdk/x/bank"           // import as blank for app wiring
 	_ "github.com/cosmos/cosmos-sdk/x/consensus"      // import as blank for app wiring
 	_ "github.com/cosmos/cosmos-sdk/x/genutil"        // import as blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/gov"            // import as blank for app wiring
 	_ "github.com/cosmos/cosmos-sdk/x/mint"           // import as blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/staking"        // import as blank for app wiring
 )
 
 var AppConfig = configurator.NewAppConfig(