+++ x/bank/autocli.go
@@ -79,9 +79,8 @@ func (am AppModule) AutoCLIOptions() *autocliv1.ModuleOptions {
 					Short:     "Query for send enabled entries",
 					Long: strings.TrimSpace(`Query for send enabled entries that have been specifically set.
 			
-			To look up one or more specific denoms, supply them as arguments to this command.
-			To look up all denoms, do not provide any arguments.
-			`,
+To look up one or more specific denoms, supply them as arguments to this command.
+To look up all denoms, do not provide any arguments.`,
 					),
 					PositionalArgs: []*autocliv1.PositionalArgDescriptor{{ProtoField: "denoms"}},
 				},
@@ -96,9 +95,9 @@ func (am AppModule) AutoCLIOptions() *autocliv1.ModuleOptions {
 					Use:       "send [from_key_or_address] [to_address] [amount]",
 					Short:     "Send funds from one account to another.",
 					Long: `Send funds from one account to another.
-			Note, the '--from' flag is ignored as it is implied from [from_key_or_address].
-			When using '--dry-run' a key name cannot be used, only a bech32 address.
-			Note: multiple coins can be send by space separated.`,
+Note, the '--from' flag is ignored as it is implied from [from_key_or_address].
+When using '--dry-run' a key name cannot be used, only a bech32 address.
+Note: multiple coins can be send by space separated.`,
 					PositionalArgs: []*autocliv1.PositionalArgDescriptor{{ProtoField: "from_address"}, {ProtoField: "to_address"}, {ProtoField: "amount", Varargs: true}},
 				},
 				{