+++ x/params/client/proposal_handler.go
@@ -1,9 +1,8 @@
 package client
 
 import (
+	govclient "cosmossdk.io/x/gov/client"
 	"cosmossdk.io/x/params/client/cli"
-
-	govclient "github.com/cosmos/cosmos-sdk/x/gov/client"
 )
 
 // ProposalHandler is the param change proposal handler.